rpi: Fix rpi images so that they build and boot.
Due to the fact that we are based on debian, not raspbian, we need to use armel NOT armhf.

Switch to using the raspbian cross compiler for the kernel though, since we already clone the tools for other uses.
Also, we don't need to use the python script anymore, the latest rpi firmware can boot from a kernel zImage, so just cp it to the proper place/name.

And lastly, add elevator=deadline for cpufreq, to the cmdline.txt for boot arguments.