Add support for building RIoTboard images.
This is a script to build images for the RIoTboard, which is a single core i.MX6 device from Embest/element14.

There won't be official images released, but the script is here for people who want to try it out.  At the moment we build off a github kernel, but 3.16 should have support for the RIoTboard in mainline.