Add an xorg.conf to force 16bit video and fbdev
Issues #49 and #47 are closely related.

Some people have issues with the beaglebone black in that modesetting doesn't
work properly and the way to workaround it is to force the fbdev driver via
xorg.conf.

And if using the onboard hdmi, or an lcd4/lcd7 you need 16bit, so set that via
the xorg.conf with a commented out section to set it to 24bit if you are using
either the bbb-exp or bbb-view.