Update all scripts with workaround for qemu
Every so often qemu will complain that a package with a lot of files in
it, typical icons, is corrupt when it is not.  Add in a little
workaround that checks that apt-get install didn't error out and if it
did, re-run to fix it.

Also add in the 3 patches for veyron, as well as a new kernel config.
Also remove my internal nameserver from veyron so people stop running
into issues when building the image.