Remove btrfs-tools
Once upon a time, btrfs was considered for the filesystem for the
chromebooks and utilite, but it ended up not panning out as well as I
would have liked.  Since we don't need it in the base images, remove it,
people can add it at any time in the future if they're actually using
btrfs on their external media.