Update banana-pi script
Updated for the new changes.  Since the kernel is old, we have to use
the older cross compiler.  Also libfdt-dev causes build issue with the
u-boot so we check if it is installed on the system, if it is, we
uninstall it, then at the end of the build, we reinstall it, otherwise
we carry on.

Also regenerate ssh keys on the first boot.  The wiggle script hasn't
been tested, so it isn't added yet.