Updated (non-working) trimslice script
Adding the trimslice script - it'll build an image, but the image won't
boot - we're moving the trimslice and riot to the kali kernels instead
of custom vendor kernels, and unfortunately, flash-kernel doesn't have a
way to detect being run under qemu so it immediately exits.  Need to go
through flash-kernel and pull out the important bits to create the
initial uboot environment, and then create a service that will install
flash-kernel onto the system.  The only reason it fails is due to the
image not being created on a trimslice.