p4wnp1: Move some things around for p4wnp1 script
The p4wnp1 script modifies the cmdline so lets make sure it exists
before it attempts to edit it.  Also because of the way the script
checks things, lets set the host name during build to something that
might resolve, and we set it to the ${hostname} variable before we
finish the image.