rpi3: fix firmware path, make a note about what's going on in the nexmon build script
We only want to build the firmware, and the nexmon firmware is... tightly integrated, so for now
We let the module error because we already have the module built in the kernel and that error occurs
after we patch up the firmware, so we get the firmware we want.  This is ugly, but it works.