rpi*: Add re4son's tft configurator, and reduce image size to 3G.
Since we resize the partitions on first boot, we can now reduce the size
of the image file to a much smaller 3GB of space.  You can still force
it to a larger size by modifying the size variable, but there's no real
need to make it larger unless you're installing a lot of software.  Keep
in mind that you need to account for both boot (~150MB) and the root
partitions when you do.

We also add re4son's new tft configurator to all the images, so once
installed, you can simply run kalipi-tft-config to set up an LCD on any
of the rpi images, as long as the script supports it.