0w: Comment out rpi-wiggle, put kalipi-config in.
So, rpi-wiggle auto running is super awesome and great.  But if you
resized the partition manually before putting it in.... rpi-wiggle would
resize the filesystem, but not correctly - it would attempt to downsize
it, or it would attempt to put the partition ending outside the bounds
of the device.  Neither of those are ideal.

So for now, we comment it out, but if anyone wants to use it instead,
they can run it manually, OR they can use the kalipi-config which should
do the right thing regardless.