odroidc2-mainline: Use fbdev, not modesetting
For some reason with the latest xorg, the modesetting video driver
causes a lot of performance issues - specifically the cursor will jump
around or freeze a lot while doing things like launching applications.
Until I've had a chance to look into why this is happening, we favor
fbdev instead.  It's not ideal, but it works.