odroidc2-mainline: Fix 1366x768 in X
When using a 1366x768 display, it's all fuzzy.  Until this is somehow
fixed in the kernel, which I don't see happening any time soon, we work
around it by testing for the resolution, and if it's set to 1366x768, we
set it to 1360x768 so we can actually use the display.

This only works in X, it would need to be set from the kernel command
line in order to work properly for the console, but I don't have a good
way to test the resolution just yet to set it that way.