rpi3*: Add hack for boot/firmware mounting
Debian's raspi3-firmware package expects /boot/firmware to exist and be
mounted.  Because we use the Raspbian way of using /boot instead, this
causes issues when attempting to dist-upgrade because it doesn't see
/boot/firmware mounted and errors out in postinst.

This is a temporary workaround, put in place until we can come up with a
better fix.