rpi3-64: Add a script to build a 64bit rpi3 image
This is a script to build a Kali Linux image for the RPi3 B, B+, and
CM3.  Nexmon appears to work, although it threw an error the first time
I ran airmon-ng start wlan0.  I tested HDMI output, and a usb
keyboard/mouse.  Ethernet works, I did not test bluetooth, but it should
work as well.

There was an issue where the timezone seemed to be set to CDT instead of
UTC, and that seems to be causing xfce to be unhappy due to a directory
somewhere having the wrong date.