rpi3: Update script to use config and comment out the sdcard patch
Patching manually, it tries to reverse the patch, so it seems to finally
be in, so lets not reverse it.  Also update the config a bit.  4D hats
won't work because they (ugh) commit object files into their repository
and they only have v6 and v7 supported versions.  Until upstream
provides a 64bit compatible object file, or more preferably, the friggin
source code to compile the object files so that it doesn't matter which
you're compiling for, we have to disable 4D hats.