rpi3-debian: Add a debian image script
Occasionally we run into issues and need to test if it's something
specific to Kali, or if it's something that affects both.  This script
builds a 32bit image that works on the rpi2 or rpi3 as the only image
out there for debian on an rpi3 (that I found) was 64bit, which in this
case didn't exhibit the issues that we're seeing.