rpi3: shared-mime-info hack and downgrade systemd
shared-mime-info doesn't seem to run properly in a chroot, so we add a
service to run it at the first boot, then disable itself so it doesn't
run every time.

also, systemd 240 is utterly broken on armhf/armel systems, so we
downgrade systemd to 239 pulled from snapshots.debian.org (but hosted by
re4son for ease of installation.)

Massive thanks to re4son for his assistance in tracking this down, and
the work around.