cuboxi: Add u-boot-menu in order to generate the extlinux.conf file
Thanks to Kartone on twitter for pointing out that the cubox-i image
wasn't actually booting.  We went through it last night and realized
it's because of the missing extlinux.conf file, so we created it
manually, but adding in the u-boot-menu package will auto generate the
file every time you upgrade the kernel.  Thanks to vagrantc for pointing
out the correct package to use on arm~