cuboxi: Make sure the cubox-i doesn't get /dev/sda3 in its cmdline
Kartone noticed that the image gets /dev/sda3 in the append line of
extlinux.conf on the image that has u-boot-menu enabled.  Unfortunately
this is a side effect of the build script running in a cross chroot.  So
we simply sed it out so that it won't be there at the end.  This won't
occur on the device itself, so the sed works well here.